Check if files deleted during updates belong to Core

+++ wcfsetup/install/files/acp/update_5.4_removeFiles.php
@@ -1,22 +1,36 @@
 <?php
+use wcf\data\package\PackageCache;
 use wcf\system\WCF;
 
 $files = [
 	'lib/system/database/table/DatabaseTableUtil.class.php',
 ];
 
-$sql = "DELETE FROM     wcf".WCF_N."_package_installation_file_log
+$sql = "SELECT  packageID
+	FROM    wcf" . WCF_N . "_package_installation_file_log
+	WHERE   filename = ?";
+$searchStatement = WCF::getDB()->prepareStatement($sql);
+
+$sql = "DELETE FROM     wcf" . WCF_N . "_package_installation_file_log
 	WHERE           packageID = ?
 	                AND filename = ?";
-$statement = WCF::getDB()->prepareStatement($sql);
+$deletionStatement = WCF::getDB()->prepareStatement($sql);
+
+$packageID = $this->installation->getPackageID();
 
 foreach ($files as $file) {
+	$searchStatement->execute([$file]);
+	$filePackageID = $searchStatement->fetchSingleColumn();
+	if ($filePackageID != $packageID) {
+		throw new \UnexpectedValueException("File '{$file}' does not belong to package '{$this->installation->getPackage()->package}' but to package '" . PackageCache::getInstance()->getPackage($filePackageID)->package . "'.");
+	}
+	
 	if (file_exists(WCF_DIR . $file)) {
 		unlink(WCF_DIR . $file);
 	}
 	
-	$statement->execute([
-		$this->installation->getPackageID(),
+	$deletionStatement->execute([
+		$packageID,
 		$file,
 	]);
 }
